Move FontTab out of ui.py
